refactor: move app stuff into it's own module
